+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComponents.java
@@ -505,8 +505,8 @@ public final class EntityComponents {
 	 */
 	public SliderBuilder slider(Attribute<Integer> attribute) {
 		AttributeDefinition<Integer> attributeDefinition = entityDefinition.attributes().definition(attribute);
-		Number minimumValue = attributeDefinition.minimumValue().orElse(null);
-		Number maximumValue = attributeDefinition.maximumValue().orElse(null);
+		Number minimumValue = attributeDefinition.minimum().orElse(null);
+		Number maximumValue = attributeDefinition.maximum().orElse(null);
 		if (minimumValue == null || maximumValue == null) {
 			throw new IllegalArgumentException("Cannot create a slider for an attribute without min and max values");
 		}
@@ -539,8 +539,8 @@ public final class EntityComponents {
 	 */
 	public NumberSpinnerBuilder<Integer> integerSpinner(Attribute<Integer> attribute) {
 		AttributeDefinition<Integer> attributeDefinition = entityDefinition.attributes().definition(attribute);
-		Number minimumValue = attributeDefinition.minimumValue().orElse(null);
-		Number maximumValue = attributeDefinition.maximumValue().orElse(null);
+		Number minimumValue = attributeDefinition.minimum().orElse(null);
+		Number maximumValue = attributeDefinition.maximum().orElse(null);
 
 		return Components.integerSpinner()
 						.minimum(minimumValue == null ? null : minimumValue.intValue())
@@ -555,8 +555,8 @@ public final class EntityComponents {
 	 */
 	public NumberSpinnerBuilder<Double> doubleSpinner(Attribute<Double> attribute) {
 		AttributeDefinition<Double> attributeDefinition = entityDefinition.attributes().definition(attribute);
-		Number minimumValue = attributeDefinition.minimumValue().orElse(null);
-		Number maximumValue = attributeDefinition.maximumValue().orElse(null);
+		Number minimumValue = attributeDefinition.minimum().orElse(null);
+		Number maximumValue = attributeDefinition.maximum().orElse(null);
 
 		return Components.doubleSpinner()
 						.minimum(minimumValue == null ? null : minimumValue.doubleValue())